Add error-first callbacks for main methods,  onComplete callback option and made the default options parameter an empty object.
Implemented callbacks with the node style of error-first callbacks to make it possibly easier to promisify with the node 'promisify' util, the onComplete option is based on GSAPs approach to end of animation callbacks. There might be some semicolons missing and some styling errors since I didnt setup eslint. No callback implementation for the methodrerun since I wasnt able to successfully use it.
 Main changes:

  - Error-first callbacks for typeString, pauseFor, deleteAll and deleteChars methods; - OnComplete option to be used as final callback, is only triggered when loop is set to false; When the options parameter is undefined its set as an empty object